requirements.txt
streamlit
torch
torchvision
torchaudio
ultralytics
easyocr
opencv-python
pillow
